Fixed load_tokens() to be more graceful
The behavior of load_tokens() has changed in a few different ways:

* load_tokens() will now automatically exclude expired tokens
if called without a list of scopes and active tokens exist
* load_tokens() will only refresh tokens for the requested scopes
given.
* load_tokens() will only raise an expired token exception given
a list of requested scopes if one of the requested scopes is
expired
* load_tokens() will only return the list of tokens requested,
and will not return other tokens or attempt to refresh them
(but it will attempt to refresh tokens for loaded scopes).