Fixed VERY sneaky server testing bug. Reduced Irony by 15%
IRONY WARNING. This bug eluded me all Saturday morning. The test
was supposed to measure the NativeClient turning off the automatic
browser if a previous login already happened. This is to prevent the
user's terminal constantly taking away control every time they ^C
out of an auth login.

It turned out this was happening correctly -- the first code handler
triggered a keyboard interrupt and set the global setting to turn off
browsers so this didn't happen. So why wasn't this working on the
remote testing server?

Well, there is also code to detect if a user is running FRL in an SSH
session and automatically turn off the browser, since browsers don't
make sense in an SSH session and only cause problems. Travis ALSO
sets env vars to state that it is a remote session. The result is that
browsers are never opened on Travis, so the mocks will never measure
that result unless tests account for this also being a remote system.

So in conclusion, FRL behaves as expected when run on a remote system.
Even if the remote system isn't a user trying to login, but Travis
trying to test code. Test passed... kindof.