fix: Packaging mistake on v0.2.4
The version on the pypi package is correct and fine, but the version
number didn't make it into the tagged version on github and was still
listed as v0.2.3. This has a small likelyhood of breaking anything,
I don't think anyone installs tags via github when PyPi exists, but
just in case I'll fix this.